(New) Histogram2D Periodic
( Not written, not compiled, not tested )
A new class that manages histograms that are meant to be used in a periodic fashion.
This is intended for signal with some inherent periodicity.
Say for example the class is instantiated with  period = 8.
Then 8 histograms will be created.
Then  when calling ::accumulate all data[i%8] points will be accumulated in the same historgram.